Update types for Sync endpoints

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonerfinancepocapi/jpa/models/NomisSyncPayload.kt
@@ -21,7 +21,7 @@ data class NomisSyncPayload(
   val timestamp: LocalDateTime,
 
   @Column(name = "transaction_id")
-  val transactionId: Int?,
+  val transactionId: Long?,
 
   @Column(name = "request_id")
   val requestId: UUID?,
