+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonerfinancepocapi/services/RequestCaptureServiceTest.kt
@@ -65,7 +65,7 @@ class RequestCaptureServiceTest {
       lastModifiedByDisplayName = null,
       offenderTransactions = listOf(
         OffenderTransaction(
-          entrySequence = 1L,
+          entrySequence = 1,
           offenderId = 1015388L,
           offenderDisplayId = "AA001AA",
           offenderBookingId = 455987L,
@@ -76,8 +76,8 @@ class RequestCaptureServiceTest {
           amount = 162.00,
           reference = null,
           generalLedgerEntries = listOf(
-            GeneralLedgerEntry(entrySequence = 1L, code = 2101, postingType = "DR", amount = 162.00),
-            GeneralLedgerEntry(entrySequence = 2L, code = 2102, postingType = "CR", amount = 162.00),
+            GeneralLedgerEntry(entrySequence = 1, code = 2101, postingType = "DR", amount = 162.00),
+            GeneralLedgerEntry(entrySequence = 2, code = 2102, postingType = "CR", amount = 162.00),
           ),
         ),
       ),
@@ -104,8 +104,8 @@ class RequestCaptureServiceTest {
       lastModifiedBy = null,
       lastModifiedByDisplayName = null,
       generalLedgerEntries = listOf(
-        GeneralLedgerEntry(entrySequence = 1L, code = 1101, postingType = "DR", amount = 50.00),
-        GeneralLedgerEntry(entrySequence = 2L, code = 2503, postingType = "CR", amount = 50.00),
+        GeneralLedgerEntry(entrySequence = 1, code = 1101, postingType = "DR", amount = 50.00),
+        GeneralLedgerEntry(entrySequence = 2, code = 2503, postingType = "CR", amount = 50.00),
       ),
     )
   }
@@ -234,7 +234,7 @@ class RequestCaptureServiceTest {
 
     @Test
     fun `should find by transactionId`() {
-      `when`(nomisSyncPayloadRepository.findByTransactionId(any<Int>())).thenReturn(listOf(dummyPayloads[0]))
+      `when`(nomisSyncPayloadRepository.findByTransactionId(any<Long>())).thenReturn(listOf(dummyPayloads[0]))
 
       val result = requestCaptureService.getCapturedNomisSyncPayloads(transactionId = 101)
